+++ GoRECIPE/gorecipe/android/local.properties
@@ -1,2 +1,5 @@
-sdk.dir=/Users/caseymaloney/Library/Android/sdk
-flutter.sdk=/Users/caseymaloney/development/flutter
+sdk.dir=/home/yaki/Android/Sdk
+flutter.sdk=/home/yaki/snap/flutter/common/flutter
+flutter.buildMode=debug
+flutter.versionName=1.0.0
+flutter.versionCode=1
