+++ GoRECIPE/gorecipe/lib/Screens/home_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 // ignore: unused_import
-import 'package:gorecipe/Screens/forgot_password.dart';
+// import 'package:gorecipe/Screens/forgot_password.dart';
 //import 'package:gorecipe/Screens/forgot_password.dart';
 import 'package:gorecipe/Screens/profile.dart';
 import 'package:gorecipe/Screens/scan_home_page.dart';
